chore(utils): reorganize utils directory
